requirements.txt
flask
flask_cors
python-dotenv
img2pdf
pillow